backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI APIs
anthropic==0.18.1
openai==1.12.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0

# Export
openpyxl==3.1.2
jinja2==3.1.3

# Redis (queue)
redis==5.0.1
celery==5.3.6
